[IMP] account_invoice_report_refund_total: all refund types coverage
Before this improvement, only those refund invoices or bills that come
from a regular invoice/bill were affected by "Total Refund" label
renaming. If e.g. a refund invoice was created by other way, it was not
applied, due to original and restricted requirements.

After this improvement, the desired functionality is applied over every
refund document, no matter if there is a reversal document is linked to
it or not.